refactoring(shell): Extract parsing and emulation logic
- Extract command parsing logic into a dedicated CommandParser class.
- Move StdOutStdErrEmulationProtocol to its own module to improve maintainability.
- Update HoneyPotShell to delegate parsing and use the extracted protocol.
- Update related commands to import the protocol from the new location.
- Fix an issue where redirection without a command was ignored.
- Add robust tests for file descriptor redirection covering edge cases.
