+++ src/cowrie/commands/sudo.py
@@ -2,7 +2,9 @@ from __future__ import annotations
 import getopt
 
 from cowrie.shell.command import HoneyPotCommand
-from cowrie.shell.honeypot import StdOutStdErrEmulationProtocol
+from cowrie.shell.std_out_std_err_emulation_protocol import (
+    StdOutStdErrEmulationProtocol,
+)
 
 commands = {}
